The ultimate fix to the most annoying emacs problem.
When started as a damon, and connected to via emacsclient, Emacs showed
wrong cursor color. It was especially annoying with zenburn theme,
because the cursor was indistinguishable from the background.

The problem was that colors from Xresources were used. So, inhibiting
Xresources fixed the problem.

One way to diagnose this issue is to look at the cursor face. If it's
state is `changed outside customize interface` then probably there is an
x-reosurces problem.

I found the solution to the problem here:   http://emacs.stackexchange.com/questions/13291/